requirements.txt
asgiref==3.9.1
Django==5.2.4
pycryptodome==3.23.0
sqlparse==0.5.3
asgiref==3.9.1
beautifulsoup4==4.13.4
Django==4.2.23
django-bootstrap-v5==1.0.11
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
pycryptodome==3.23.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
